requirements.txt
gradio==5.33.2
langchain==0.3.25
langchain-core==0.3.65
langchain-ollama==0.3.3
langchain-community==0.3.25
langsmith==0.3.45
ollama==0.5.1
scikit-learn==1.7.0
pandas==2.3.0
numpy==2.2.6
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.1.2
pdfplumber==0.11.7
docx2txt==0.9
PyMuPDF==1.26.1
python-dotenv==1.1.0
joblib==1.5.1
chromadb==1.0.12